clean up & move logic to separate module
Removes cyclic dependency and redundant code
