+++ mythril/laser/smt/__init__.py
@@ -1,5 +1,8 @@
 from mythril.laser.smt.bitvec import (
     BitVec,
+)
+
+from mythril.laser.smt.bitvec_helper import (
     If,
     UGT,
     ULT,
@@ -15,6 +18,7 @@ from mythril.laser.smt.bitvec import (
     BVSubNoUnderflow,
     LShR,
 )
+
 from mythril.laser.smt.bitvecfunc import BitVecFunc
 from mythril.laser.smt.expression import Expression, simplify
 from mythril.laser.smt.bool import Bool, is_true, is_false, Or, Not, And
